Make avails filter a reusable component
