+++ apps/catalog/catalog/src/app/marketplace/movie/search/search.component.html
@@ -256,100 +256,7 @@
                 </span>
               </mat-expansion-panel> -->
             </ng-container>
-            <ng-container [formGroup]="availsForm">
-              <span class="mat-subheading-1">AVAILS</span>
-              <p class="mat-caption">You need to fill in Terms, Territory and Media before applying the Avails filter</p>
-              <!-- Availabilities -->
-              <mat-expansion-panel>
-                <mat-expansion-panel-header>
-                  <mat-panel-title test-id="availabilities-panel">Terms</mat-panel-title>
-                </mat-expansion-panel-header>
-                <ng-container formGroupName="terms">
-                  <mat-form-field appearance="outline">
-                    <mat-label>Start Date</mat-label>
-                    <input matInput formControlName="from" [matDatepicker]="pickerAvailableFrom" placeholder="Start Date" />
-                    <mat-error *ngIf="availsForm.get('terms').get('from').hasError('min')">
-                      You only can choose the current year and above.
-                    </mat-error>
-                    <mat-datepicker-toggle test-id="datepicker-from" matSuffix [for]="pickerAvailableFrom">
-                    </mat-datepicker-toggle>
-                    <mat-datepicker #pickerAvailableFrom startView="multi-year"></mat-datepicker>
-                  </mat-form-field>
-                  <mat-form-field appearance="outline">
-                    <mat-label>End Date</mat-label>
-                    <input matInput formControlName="to" [matDatepicker]="pickerAvailableTo" placeholder="End Date" />
-                    <mat-datepicker-toggle test-id="datepicker-to" matSuffix [for]="pickerAvailableTo">
-                    </mat-datepicker-toggle>
-                    <mat-datepicker #pickerAvailableTo startView="multi-year" startAt="2020"></mat-datepicker>
-                  </mat-form-field>
-                </ng-container>
-                <mat-error *ngIf="availsForm.get('terms').hasError('invalidRange')">
-                  The value <i>To</i> can't be smaller than the value <i>From</i>!
-                </mat-error>
-              </mat-expansion-panel>
-
-              <!-- Territories -->
-              <mat-expansion-panel>
-                <mat-expansion-panel-header>
-                  <mat-panel-title test-id="territories-panel"> Territory </mat-panel-title>
-                </mat-expansion-panel-header>
-                <mat-form-field appearance="outline">
-                  <mat-chip-list #territoryChipList>
-                    <mat-chip *ngFor="let territory of selectedMovieTerritories; let i = index" [selectable]="selectableTerritory"
-                      [removable]="!availsForm.value.isActive" (removed)="removeTerritory(territory, i)">
-                      {{ territory }}
-                      <mat-icon matChipRemove *ngIf="!availsForm.value.isActive">cancel</mat-icon>
-                    </mat-chip>
-                    <input test-id="territories-panel-input" placeholder="Territories available" #territoryInput
-                      [matAutocomplete]="territoryMatAutoComplete" [matChipInputFor]="territoryChipList" [formControl]="territoryControl" />
-                  </mat-chip-list>
-                  <mat-autocomplete #territoryMatAutoComplete="matAutocomplete" (optionSelected)="selectedTerritory($event)">
-                    <mat-option *ngFor="let territory of territoriesFilter$ | async" [value]="territory">
-                      {{ territory }}
-                    </mat-option>
-                  </mat-autocomplete>
-                  <mat-hint>Please choose a territory out of the drop down menu!</mat-hint>
-                </mat-form-field>
-              </mat-expansion-panel>
-
-              <!-- Media -->
-              <mat-expansion-panel>
-                <mat-expansion-panel-header>
-                  <mat-panel-title test-id="mandate-medias-panel">Media</mat-panel-title>
-                </mat-expansion-panel-header>
-                <div fxLayout="column" fxLayoutAlign="start space-between" class="filter-checkboxes-media">
-                  <mat-checkbox *ngFor="let media of movieMedias" (change)="checkMedia(media)" [disabled]="availsForm.value.isActive">
-                    {{  media | translateSlug: 'MEDIAS' }}
-                  </mat-checkbox>
-                </div>
-              </mat-expansion-panel>
-
-              <!-- Exclusivity -->
-              <mat-expansion-panel>
-                <mat-expansion-panel-header>
-                  <mat-panel-title>Exclusivity</mat-panel-title>
-                </mat-expansion-panel-header>
-                <article fxLayout="row" fxLayoutGap="10px">
-                  <span>No</span>
-                  <mat-slide-toggle formControlName="exclusivity" color="primary"></mat-slide-toggle>
-                  <span>Yes</span>
-                </article>
-              </mat-expansion-panel>
-              <article fxLayoutAlign="center center" class="filter-buttons">
-                <button
-                  *ngIf="!availsForm.value.isActive"
-                  mat-stroked-button color="primary"
-                  (click)="applyAvailsFilter()">
-                  Apply Avails Filter
-                </button>
-                <button
-                  *ngIf="availsForm.value.isActive"
-                  mat-stroked-button color="warn"
-                  (click)="deactivateAvailsFilter()">
-                  Deactivate Avails Filter
-                </button>
-              </article>
-            </ng-container>
+            <catalog-avails-filter [availsForm]="availsForm"></catalog-avails-filter>
         </mat-accordion>
       </mat-card-content>
     </mat-card>
